JVMTIL-10 残業算つくった - クラスを分割した
