+++ java-projects/miscellaneous/src/main/java/com/example/over/OvertimeWork.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2017 Shinya Mochida
+ * 
+ * Licensed under the Apache License,Version2.0(the"License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing,software
+ * Distributed under the License is distributed on an"AS IS"BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.example.over;
+
+import java.time.Month;
+
+public interface OvertimeWork {
+    Month getMonth();
+
+    int getHours();
+
+    default boolean is45HoursOver() {
+        return getHours() >= 45;
+    }
+
+    OvertimeWork nextMonth(int hour);
+}
